+++ src/TestTrackingDiagrams.Extensions.EventHubs/EventHubsOperation.cs
@@ -1,0 +1,17 @@
+namespace TestTrackingDiagrams.Extensions.EventHubs;
+
+public enum EventHubsOperation
+{
+    Send,
+    SendBatch,
+    CreateBatch,
+    ReadEvents,
+    ReadEventsFromPartition,
+    GetPartitionIds,
+    GetEventHubProperties,
+    GetPartitionProperties,
+    StartProcessing,
+    StopProcessing,
+    ProcessEvent,
+    Other
+}