+++ src/TestTrackingDiagrams.Extensions.EventHubs/EventHubsTrackingVerbosity.cs
@@ -1,0 +1,8 @@
+namespace TestTrackingDiagrams.Extensions.EventHubs;
+
+public enum EventHubsTrackingVerbosity
+{
+    Raw,
+    Detailed,
+    Summarised
+}